Check Condition on System.Text.Json.JsonIgnoreAttribute
I'm typically adding
`[JsonIgnore(Condition = JsonIgnoreCondition.Never)]`
to my classes, to ensure that serialized json has the format I want,
even if someone changes the global `JsonIgnoreCondition`.

This means that the existance of `JsonIgnore` is not enough, but you
need to check that `Condition == JsonIgnoreCondition.Always`.

The Docs for `[JsonIgnore]` states that `Always` is the default condition
and the other conditions does not indicate that the property should be
ignored from schema.
https://docs.microsoft.com/en-us/dotnet/api/system.text.json.serialization.jsonignoreattribute.condition?view=net-6.0